+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.3.0 - 2017-09-22
+
+### Changed
+
+* Updated `container-interop/service-provider` to `0.4.0`
+
+### Removed
+
+* Removed the delegate lookup feature
+
 ## 0.2.1 - 2017-04-13
 
 ### Fixed